#27 Move enums and their option details to specific files
