+++ frontend/src/components/skills/DraggableSkillCard.tsx
@@ -1,7 +1,7 @@
 import SkillCard, { SkillCardProps } from './SkillCard';
 
 import { useDrag } from 'react-dnd';
-import { ItemType } from '../../types/enums/SkillEnums';
+import { ItemType } from '../../types/enums/ItemType';
 
 export default function DraggableSkillCard({ skill, ...skillCardProps }: SkillCardProps) {
     const [, dragRef] = useDrag(() => ({
