+++ frontend/src/types/enums/ItemType.ts
